requirements.txt
torch==2.3.0
torchvision==0.18.0
transformers==4.40.2
accelerate==0.30.1
peft==0.11.1
datasets==2.19.1
pandas
pillow==10.3.0
flash-attn==2.5.8
numpy
bitsandbytes